media: 64 bit compile issues
- change internal sized types to use stdint.h
- printf & scanf formats
- size_t or unsigned int for iterators

+++ include/media/stagefright/FileSource.h
@@ -55,7 +55,7 @@ private:
     sp<DecryptHandle> mDecryptHandle;
     DrmManagerClient *mDrmManagerClient;
     int64_t mDrmBufOffset;
-    int64_t mDrmBufSize;
+    size_t mDrmBufSize;
     unsigned char *mDrmBuf;
 
     ssize_t readAtDRM(off64_t offset, void *data, size_t size);
